requirements.txt
PyGithub==2.3.0
pyyaml==6.0.1
